requirements.txt
beautifulsoup4==4.10.0
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
decorator==5.1.0
future==0.18.2
geocoder==1.38.1
idna==3.2
jsonlib-python3==1.6.1
loc==0.0.2
MouseInfo==0.1.3
Pillow~=10.4.0
psutil~=6.0.0
PyAudio
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3==2.90
pytube==11.0.1
pytweening==1.0.4

ratelim==0.1.6
requests~=2.32.3
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
urllib3==1.26.7
wikipedia==1.4.0

opencv-python~=4.10.0.84
pytesseract~=0.3.10
numpy~=1.26.4
playsound~=1.3.0
openai~=1.35.14
dlib~=19.24.4
transformers~=4.42.4
imutils~=0.5.4
nylas~=6.3.0
httplib2~=0.22.0
apiclient~=1.0.4
google-api-python-client~=2.139.0
oauth2client~=4.1.3